Improve Jump Statement Resolution Handling
Add function "symbol_table_process_control" which resolves jump
statements based on the symbol of the statement which owns the table.
The goal is to deduplicate the code which does this in the type checker
and also to shift responsibility away from the type checker, which has
grown to do a lot more than just type checking.