Implement New Jump Checking Functions
Implement jump checking stack functions in state.c. Modify simplestack.h
to be useful for this purpose, since it was unused and the information
required by break and continue statements is too simple for badlib's
linked list to be a good solution. More planning on ways to simplify
compiler or fix broken/unimplemented functionality.