Improve Symbol Table Printing
Symbol table print function now indents based on the depth of the scope
in which the symbols were declared, sorts symbols based on the location
that they were declared in and the location of their parent scope, and
prints more detailed information about their scope. For example, if
their scope belongs to a `for` loop, `while` loop, or other statement,
the scope is listed as belonging to that statement.