Improve/Change Symbol Table Output, Minor Bugfix
Print symbols for functions before printing the contents of their
respective function scope, which is achieved by setting the location
of the function's symbol to the location of the identifier, rather than
the location of the first specifier. Print symbol tables for blocks
which are children of `case`, `default`, and labels. Modify format
string for symbol table output. Fix big where functions without
parameters would segfault because their symbol table was `NULL`.