Implement `types_equivalent` and Helper Functions
Implement `types_equivalent` and helper functions for checking auxiliary
specifiers and parameter equivalence. Declaration specifiers are not
quite compared correctly because all we have to go off of are the width,
alignment, base type, and a set of flags indicating which declaration
specifiers were written. This is enough to deduce the implied
specifiers, but would require much more work and the current solution
should work fine for the x64 System V ABI.