Single Pass Generation of Assignments and Calls
Add single pass intermediate language generation for function calls and
the assignment operator. Update `save_registers` and `restore_registers`
so that they take as an argument a list iterator, which is manipulated
to point at the last push/pop instruction added to the list. Correct
stack save/restore order for function call operator.