Fix Code Emitted for Increment/Decrement
Fix code emitted for increment and decrement. Many extra instructions
must be emitted since the generator expects every instruction to have at
least one register operand (so that it can infer width), and expects the
destination operand of the last instruction emitted for a node to
contain the result of that node, if applicable.