Proper Condition Conversion Conversion for Switch
Track type of the controlling expression for switch statements. Emit
conversions for the controlling experssion of switch statements and for
the values of case statements. Change error printing code for the
"expected some type" class of errors so that non-integer switch
controlling expressions can be reported in the new code.