Fix Pointer Arithmetic in Constexprs
Fix constexpr evaluator not multiplying/dividing integral operands and
results by the member width when doing mixed integral and pointer
addition and subtraction, and when subtracting pointers. Remove
erroneous attempt to evaluate subscripts when the index contains an
address. Rename the `offset` component of address constant values to
`disp` to match elsewhere in the compiler.