Improve Handling of Specially Created Type Structs
Improve handling of resources created by expression nodes which copy and
modify types. Rather than allocating resources every time the type
checker needs to add a a pointer auxspec or change an existing auxspec
to a pointer, there now constants used for this purpose. When formatting
a type as a string, print type qualifiers for pointers and basic types.