Change Insructions and Order for Function Calls
Change order in which instructions are emitted during function calls and
add additional instructions to the sequence. The return value of a
function is now placed onto the stack temporarily so that the volatile
registers can be safely restored. Emit constant expression load code for
the function being called so that it is before all argument
instructions; previously it would be located after non-constant
expression arguments.