Allow Void Return Statements to Have Expressions
Allow functions returning `void` to have expressions in their return
statements, so long as the result of the expression is cast to `void`.
Add a little bit of error checking to `scalar_conversions` to ensure
that the given types are valid. Reuse `skip_asm` flag to prevent
printing of assembly in order to check for errors.