Miscellaneous Fixes
- fix `scalar_conversions`; make it correctly reject invalid source and
destination types, and handle enums properly.
- fix `translate_logical_not` not setting instruction iterators correctly.
- fix `translate_indirection` setting result width incorrectly and
emitting the incorrect instruction when the result type is aggregate.
- fix jump and case ids being improperly set and compared.
- fix order in which case ids are assigned to case statements.
- add `type_is_record` function for convenience.
- fix double free bug caused by `validate_conditional`. Whenever the
result type is a pointer, the pointer declarator must always be copied
in case the result is a common qualified pointer type; determining
whether or not this is the case at destruction time is much harder than
always allocating a little memory for the pointer type.