Fix Some `tchk_decl` TODOs
- Remove redundant check from `validate_tag_typespec`.
- Add check for deduced length when completing array types.
- Copy type information in place when completing deduced array types.
- Move some TODOs out of `tchk_decl` and into the source files where
the changes talked about need to take place.
- Remove some type checking code from `tchk_decl`. It will be relocated
to `init`.
- Change error codes to asserts in `define_function`. These situations
indicate a bad bug in the compiler and it should just die.
- Change `assert`s to `abort()` in `validate_tag_decl`. These conditions
should cause the compiler to fail even in release builds.
- Use `UNWRAP` macro in `finalize_tag_def` when tag node is an error.
- Ensure that the scope being popped in `finalize_tag_def` belongs to the
aggregate being finalized in case an error occurred before the scope
could be pushed.
- Add comment that tells the reader where code is emitted for
initiializers since it might not be in the expected place.
- Remove unnecessary TODOs