Fix Some `tchk_expr` TODOs
Remove some TODOs that were unnecessary or already fixed. Simplify
partial copy of pointer type in `validate_conditional` when one operand
is pointer and the other is constexpr zero, and change an `assert` to a
call to `abort`. Add note that functions and arrays should be considered
non-assignable lvalues in case I forget.