Fix Some `asmgen` TODOs
- Separate `translate_local_init` into three separate functions.
- Reject `a += b` and `a -= b` in `tchk_expr` where `a` is of integral type and
`b` is of pointer type.
- Remove code from `assign_add` which handled `a += b` where `a` is of integral
type and `b` is of pointer type, as this is not semanticall valid.
- Add helper function for translating addition in the case that one operand is
of pointer type and the other is of integral type.
- Add (void) pointer type constant.
- Fix `translate_indirection`; the only instructions it should need to emit are
those generated by `scalar_conversions` when it turns lvalues into rvalues.
- Fix `translate_addrof`; it does not need to emit any instructions. The
semantically important thing the address-of operator does from the compiler's
perspective is clear the lvalue attribute from the syntax tree.
- Change the way `assign_add` loads result for use by parent expression to
make reasoning about persistence flags easier.
- Use `r10` instead of `rax` when an arbitrary volatile register is needed since
it is never used by the function call sequence.
- Always widen scalar function call arguments to be eight bytes wide to make it
easier to load and store them.
- Change `assign_stack_space` signature to take a type as an argument and return
the displacement of the stack region assigned to reduce the number of dummy
values needed when calling it.
- Create a fake type and tag used to assign stack space for spilled register
varargs.
- Reorder members in `TagValue`'s `data` union member so that its `members`
member may be initialized.
- Create separate stack regions for hidden function parameters and temporary
storage for function return values to prevent them from clobbering each other.
- Reorder creation of instructions in `translate_va_start` to mirror the order
in which they are emitted.
- Emit a no-op when a block statement contains no statements which emit
instructions.
- Use `maybe_load_cexpr` to load case statement constants and convert these
constants to the same width as the control expression.
- Add comments which note the behavior of `liter_push_back` and
`liter_push_front` when the output parameter points to the input parameter.
- Separate `translate_local_init` into multiple functions.
- Add constants for the displacements of the hidden parameter and function
return value stack storage regions and for the initial stack window size.