Improve Symbol/Tag/Specifier Naming (Hopefully)
Improve/change the names of tags, objects and functions, which should
hopefully improve clarity. A list of the naming changes made follows,
which should cover most if not all of the changes:

- `SymbolValue` -> `Symbol`
- `TagValue` -> `Tag`
- `symbol_value`, `symval` -> `symbol`
- `tag_value`, `tagval` -> `tag`
- `TableType` -> `TableKind`
- `TagType` -> `TagKind`
- `tag_type`, `table_type` -> `kind`
- `typespec`, `spec` -> `decl_spec`
- `typespecs`, `typespec_list` -> `decl_specs`
- `tag_name_node`, `ident_node` -> `ident`
- `tag_name`, `tagname` -> `id_str`
- `symbol_name`, `sym_name`, `symname` -> `ident`
- `tag_type_node` -> `tag_spec`
- `type_id` -> `typedef_name`

The `symbol` member of `ASTree` is also renamed to `tok_kind`, to avoid
confusion with `symbol`, which is used for objects of the `Symbol`
type.