Change Internal Structure of `Symbol`
Replace `flags` member of `Symbol` structure with separate members for
storage class, linkage, and a field indicating whether or not a symbol
has been defined. Change location member to be a pointer; symbols should
always have a location associated with a node in the syntax tree which
declared them (or a constant value for internal symbols), so there's no
reason to copy the location information. Remove some other useless
fields from the structure definition.