Handle Declaration of Block Extern Symbols
Handle declarations of symbols with the `extern` keyword at block and
function scope properly. Remove fake linkage types for invalid linkage
and inherited symbols and replace with a new member that indicates
whether a symbol is inherited, hidden, and does or does not have a
definition, all of which are mutually exclusive conditions. Add
functions which handle array completions and adding prototypes to
functions, which as a consequence greatly simplifies
`validate_declaration`. Combine functions for determining the linkage
and storage class of a symbol into a single function.