Improve Handling of DIV/IDIV/MUL/IMUL
Make the compiler emit a little song-and-dance every time it has to
perform a multiplication, division, or remainder operation. This must be
done because the register allocator is bad and has no concept of
implicit operands or how to deal with instructions that have already had
a real register assigned to them. Add instructions for sign extending
`rax` into `rdx` for use with `div` and `idiv`. Improve some test files
since the compiler can handle standard library headers now so that
program output can be compared to output when compiled with gcc alone.