Simplify Syntax and Lexical Error Handling
Stop lexer/parser when a syntax or lexical error is encountered and
terminate compilation gracefully. Remove `%destructor%` from yacc input
file and handle freeing discarded tokens manually. Free type information
from struct, union, enum and declarations specifier list tokens if it is
present, which will be the case if an error occurs in the middle of a
declaration. Comment out uses of `skip_type_check` since it is not
currently used. Add some small test cases for lexical and syntax errors.