Use Alternative Error Mechanism for Declarations
Remove uses of abstract syntax tree based error handling mechanism from
portions of the semantic analyzer that handle declarations. Add
`type_validate`, which determines if a declarator's type is valid; in
particular, it catches declarations of objects with void type,
declarations of arrays of functions and declarations of arrays of
functions or arrays of elements with incomplete type. Make `tag_print`
visible outside of its translation unit.