Return Status Codes When Analyzing Initializers
Make `traverse_initializer` and all internally linked function it
depends on return integer status codes indicating success or failure,
as opposed to `void` or `ASTree*`. This status code is not currently
checked; in the future it will be used if the code generator needs to
stop early, before the next parser rule is matched, as the result of a
semantic error.