Fix `translate_params` Using Incorrect Dirdecl
Retrieve first child of declarator syntax tree node for function
definitions rather than the last child. This node should be the direct
declarator for the function. Use `type_param_count` to determine the
number of function parameters rather than the number of children of the
function direct declarator, which simplifies the logic of determining
how many parameters a function has due to `void` and `...` tokens also
potentially being children of this node.