Use Iterators Less for Instruction Emission
Use fewer iterator parameters to determine where instructions should be
inserted, in particular in bulk register/memory movement functions, when
emitting function call sequences and when emitting for initializers.
This significantly changes the way function calls look; the stack is
adjusted for parameters with a single `sub` instruction, and arguments
are no longer emitted in reverse order.