Fix Returning of Aggregates from Functions
Fix register clobbering when returing aggregate values from functions.
When returning small aggregates, live values in `rax` and `rdx` could be
clobbered when moving the return value into these registers. When
returning large aggregates on the stack, `rax` was used to hold the
destination address, which would clobber the source address if it was
also held in `rax`.