Make Function Prologue and Epilogue More Normal
Make code generated for the function prologue and epilogue look more
normal; that is, it begins with `push %rbp ; mov %rsp, %rbp` and pushes
the rest of the callee-saved registers afterwards. This is mirrored in
the epilogue. Adjust fixed displacements, stack allocation, and window
size padding to account for the changes this makes to the value of the
stack pointer.