Divide Result of Pointer Differences
Divide the result by element width when subtracting two pointers. This
is implemented with one or both of an arithmetic right shift and a
multiplication by the integral multiplicative inverse of the element
width mod 2^64, found using the extended Euclidean algorithm. Add a test
which makes sure that pointer differences are being calculated
correctly.