Fix Operand Mismatch in `convert_scalar`
Fix `convert_scalar` emitting `mov` instructions with mismatched
operand widths when handling "conversions" from 32-bit to 64-bit values
without sign extension. Whenever the lower 32 bits of a register are set
on AMD64, the upper 32 bits should be zeroed, so all `convert_scalar`
needs to do in these situations is make sure the instruction it emits
has a destination with the correct operand width.