Fix Incorrect Width of String Literals
Fix length of string literals being calculated incorrectly using
`strlen`. Escape sequences only occupy a single byte in memory, so the
length of strings containing escape sequences must be calculated
excluding all but one of the characters of the sequence. Fix other
locations in the code using the length of the literal's lexical
information to determine array size, rather than the type information
attached to the literal's token.