Fix Out-of-Order Instruction Emission for Calls
Fix instructions for function calls being emitted in incorrect order as
a result of conversions not being applied to the designator until after
the expression for the first argument had been parsed. Conversion
instructions would be emitted after the first argument's instructions,
causing the first argument's instructions to be "orphaned" outside the
range of the function call's range of instructions.