Fix Bugs Preventing Compilation of Flex/Bison Files
Emit `nop` instructions for type names at block and function scope to
satisfy the expectation that every declaration and statement at block
and function scope emit at least one instruction.

Append `va_arg` instructions after the type name's single `nop`
instruction so that the result of the operation is held in the last
instruction emitted for the expression.

Allow partial types created for declaration specifier lists to include
direct declarators. This is possible through the use of typedef names.

Perform constant expression conversions on the operand to `va_arg`,
`va_start`, and `va_end`. None of these expressions can be constant
expressions, so their argument must be converted.

Define the `fileno` function in `lyutils` since the `-std=c90`
compiler/preprocessor option results in this function not being defined
in `stdio`.

All of these changes allow the compiler to accept source files generated
by GNU Flex and Bison as valid input, with one exception. A single line
in the generated parser attempts to use identifiers in an intitializer
list, which, if I am reading the standard correctly, is not allowed in
C90, and would require significant changes to accommodate. For the time
being, it is less of a hassle to just change this line manually when
testing self-compilation.