Replace HashTable Implementation in `regalloc`
Use the new macro-based hashtable implementation in register allocator
to help improve type safety.

Add a function to clear all entries from the macro-based hastable
implementation, which is required by the register allocator.

Add a function to count the number of entries in the hashtable using the
neighborhood bitsets. This function is not necessary since the hashtable
already tracks the number of keys inserted, but is useful to check to
make sure that the table is functioning correctly.