Replace Badlib Linked List for Instructions
Change instruction data structure so that it may be treated as a
circular doubly-linked list independently of the badlib linked list.
This is particularly invasive, requiring nearly every file to be
modified, but makes adding instructions significantly easier, removes
the need to worry about managing iterators, and increases the compiler's
speed by a surprising amount.

Append new instructions to the syntax tree node they are created by,
rather than a global instruction list. Instruction lists are
concatenated when the parent of the node they belong to is translated.
Exceptions include directives for block- and function-scope objects with
static storage class and directives for string literals, which are
appended directly to the root node's instruction list.

Change parsing of definitions slightly to allow them to be inserted into
the symbol table before their initializer is parsed without emitting
directives twice.

Add the `ERASED` pseudo-instruction which is used to replace the `zero`
directive when an object with static storage class is defined separately
from its declaration. The directive cannot simply be removed because,
for debugging purposes, syntax tree nodes retain knowledge of which
instructions "belong" to them. References to the deleted instruction by
these nodes would also have to be changed, but may be inaccessible when
the deletion occurs.

Handle `MODE_SYMBOL` operands in register allocator.