Unify Dynamically Allocated Array Management
Replace all usages of dynamically allocated arrays with a unified
macro-based interface to help bugs related to memory management and to
reduce the number of lines of code dedicated to resizing arrays.

Remove simple stack macro generator, as the new array management macros
are more than adequate for the same task.