Streamline Hashtable Functions
Use the bitwise and operator, `&`, instead of the remainder operator,
`%`, when restricting hashtable indices since it much faster and the
hashtable size is guaranteed to be a power of two.

Remove thunk function used when calculating hashes, since its previous
behavior was not suitable for the changes made to the hash table's grow
function and its new behavior is simple enough that it does not warrant
a separate function.

Reduce complexity and increase speed of table resizing operation by
exploiting the guaranteed size of the hash table and properties of a
hopscotch hash table. The new implementation performs the resize
operation in place, checks that keys move before making modifications to
the table, and is not recursive, as the resize operation is guaranteed
to succeed assuming memory allocation does not fail.

These changes are mostly to increase the performance of self-compiled
code, as this compiler does not perform any optimization.