Report Invalid Uses of Incomplete/Function Types
Check for declarations of objects of incomplete or function type in
places where they are not allowed. This commit adds checks for and
terminates compilation if incomplete types are used for:

1. parameter types in function definitions
2. return types in function definitions
3. member types in record definitions
4. declarations at block or function scope
5. definitions at any scope, excluding deduced size arrays

Additionally, compilation is terminated if an object of `void` type is
declared. Declarations with the `typedef` storage class specifier are
excluded from checks for incomplete type.

Check for and terminate compilation if a record member is declared with
function type.

Check for and terminate compilation if a function is defined having
achieved function type through a typedef. In other words, if a function
definition does not have a parameter list of some form, it is not valid.

Add tests which cover the above cases.