Uniform Tree Structure for Tag Definitions
Generate semantic tree nodes for unnamed structs, unions, and enums and
store their unique names in the generated string set. This removes the
need to check for identifier nodes as the structure of tag declarations
is always the same, and also removes the need for scopes to free unique
names.