Identify Lvalues Without Using Flags
Remove `ATTR_EXPR_LVAL` attribute flag. The flag was never necessary;
you can determine whether or not a node is an lvalue and whether or not
it is modifyable based on its token, type, and whether or not its value
is a compile time constant.

Add `astree_is_lvalue`, which determines whether or not a syntax tree
node represents a (modifyable) lvalue based on the criteria listed
above.

Remove `ATTR_STMT_DEFAULT` attribute flag. This flag was unused; the
global state tracks whether or not a switch statement has a default
label.

Rename the syntax tree's attributes member, enum type, and enum
definitions to reflect that its only use now is to differentiate
constant expressions and that it no longer represents bit flags, but
actual enumeration constants.

Remove `type_is_object` and `symbol_is_lvalue` functions. They are no
longer necessary because of the new way lvalues are handled and their
names and purposes were weird.