Allow Any Type to be Cast to `void`
Fix type checking of cast expressions to allow expressions of any type
to be cast to `void`. Rewrite existing checks for casts between scalar
types to be more specific and correct. Check for and report errors when
source and/or destination type of cast expressions are incomplete. Amend
`nested_tags` test to include a cast from an aggregate type to `void`.